lighttpd: prefer lighttpd strong TLS defaults
Since lighttpd 1.4.68, released Jan 2023 (over a year ago), TLS defaults
in lighttpd are at least as strong as cipherlist.eu recommendations.

For those setting up new lighttpd instances and referencing
cipherlist.eu for new configs, the lighttpd TLS defaults are strongly
recommended, as the lighttpd TLS defaults are periodically reviewed
and updated to be stronger.

Case in point, the next release of lighttpd will be lighttpd 1.4.75
and will include TLS defaults that are stronger than those currently
recommended on cipherlist.eu.

Anyone using older versions of lighttpd *and* referencing cipherlist.eu
for new configs for very old releases of lighttpd -- or very old
releases of other servers -- probably ought to be directed to
https://ssl-config.mozilla.org/ for configuring older server versions.